Initial commit
This is the beginning of the Griffin project that aims to be a
sister framework to Elixir Phoenix. While Phoenix might be more
capable of building complex web applications, there is space for a
simpler yet robust tool for building websites.
The ambitious goal of this project is to become the de facto tool
for Erlang and Elixir users to build static sites. To help in this
goal we start by providing a tool capable of converting Markdown
files into HTML with basic layout functionality. After this core
feature is fully implemented, we will add support for plugins and
themes, which we consider to be crucial for tools like this.
Griffin's bootstrapping code and accompanying mix tasks lean
heavily on Phoenix, and they share the same license.