Fix layout names being stored with file extensions (#9)
* Fix layout names in templates

This commit fixes an issue with layout names for regular and partial
layouts that that forced users to refer to partials using quoted
atoms (e.g. `@partials."footer.html"`) due to the way we were parsing
layout names. For files that use multiple extensions like `.html.eex`,
it is now possible to refer to them using their name without any
extension.

In content, pages can now also refer to the layout just as `layout_name`
instead of including the file extensions of the layout.

* update plug_live_reload dependency